add confirmation to deleting notes

+++ src/components/MessageGrid.svelte
@@ -59,12 +59,15 @@
     let accessKey = Cookie.get('awsAccessKey');
     let secretKey = Cookie.get('awsSecretKey');
     let client = documentClient(accessKey, secretKey);
-    let success = await Promise.all(
-      Array.from($tiles_checked).map(async (created) => {
-        return await deleteSnippet(client, created);
-      })
-    );
-    window.location.reload();
+    let ok = window.confirm("Are you sure you want to delete notes?");
+    if (ok) {
+      let success = await Promise.all(
+        Array.from($tiles_checked).map(async (created) => {
+          return await deleteSnippet(client, created);
+        })
+      );
+      window.location.reload();
+    }
   }
 
 </script>
